requirements.txt
pycryptodomex
requests
passlib
beautifulsoup4
lxml
